requirements.txt
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.0.0
numpy>=1.20.0
tqdm>=4.50.0
